requirements.txt
Django==2.0.3
django-discover-runner==1.0
django-enumfields==0.10.0
django-role-permissions==2.2.0
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
six==1.11.0
